+++ AEPAssurance/Source/AssurancePresentation.swift
@@ -15,24 +15,24 @@ import Foundation
 ///
 /// An Assurance Session Presentation protocol which represents common presentation logic needed for an assurance presentation
 ///
-protocol AssurancePresentation {
-    ///
-    /// Called when a session connection has been successfully established.
-    ///
-    func sessionConnected()
-    
-    ///
-    /// Called when a session has been disconnected
-    ///
-    func sessionDisconnected()
-    
-    ///
-    /// Called when a session connection error has occurred
-    ///
-    /// - Parameter: the error `AssuranceConnectionError` which occurred
-    ///
-    func sessionConnectionError(error: AssuranceConnectionError)
-}
+//protocol AssurancePresentation {
+//    ///
+//    /// Called when a session connection has been successfully established.
+//    ///
+//    func sessionConnected()
+//    
+//    ///
+//    /// Called when a session has been disconnected
+//    ///
+//    func sessionDisconnected()
+//    
+//    ///
+//    /// Called when a session connection error has occurred
+//    ///
+//    /// - Parameter: the error `AssuranceConnectionError` which occurred
+//    ///
+//    func sessionConnectionError(error: AssuranceConnectionError)
+//}
