+++ TestApp/ContentView.swift
@@ -86,7 +86,11 @@ struct AssuranceCard: View {
                     }
                 }, label: {
                     Text("Connect")
-                }).buttonStyle(YellowButtonStyle()).padding()
+                }).buttonStyle(YellowButtonStyle()).padding().onReceive(NotificationCenter.default.publisher(for: .deviceDidShakeNotification)) { _ in
+                    #if DEBUG
+                    Assurance.startSession()
+                    #endif
+                }
             }
         }
     }
@@ -249,3 +253,15 @@ struct BigEventsCard: View {
         }
     }
 }
+#if DEBUG
+extension NSNotification.Name {
+    public static let deviceDidShakeNotification = NSNotification.Name("MyDeviceDidShakeNotification")
+}
+
+extension UIWindow {
+    open override func motionEnded(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
+        super.motionEnded(motion, with: event)
+        NotificationCenter.default.post(name: .deviceDidShakeNotification, object: event)
+    }
+}
+#endif
